model of mvvm
